+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project started near the end of the 2023, due my necessity to improve my ab
 I tried - same time ago - to use electron, but at last eventually I decided to try it again in python, and I am seeing good results in terms of lightness and practicality of development.
 
 # Versions
+## pengui-0.2.1
+* added calamares --install, calamares --remove, status and cuckoo (disabled);
+
 ## pengui-0.1.9
 * penGUI now is used always with normal user, no root:
 * before to open a Terminal, penGUI will show a message box, to accept or nome.